requirements.txt
ccxt==4.2.67
pandas==2.2.1
websockets==12.0
python-dotenv==1.0.1

############################# only on LINUX
# TA-Lib Installation:
# curl -SLO http://downloads.sourceforge.net/project/ta-lib/ta-lib/0.4.0/ta-lib-0.4.0-src.tar.gz
# tar -xzf ta-lib-0.4.0-src.tar.gz
# cd ta-lib/
# ./configure --prefix=/usr
# make
# sudo make install
# TA-Lib==0.4.28 
#############################

pandas-ta==0.3.14b